+++ CHANGELOG.md
@@ -73,6 +73,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Explorer, CSV Explorer
 - Deactivate hapi fhir client and hibernate logs (configurable)
 - FHIR Terminology Server
+- Location String and Message for validation
+- Parameter LOG_FHIR_VALIDATION
 
 ### Changed
 - Rename Conversion Template as Converter Template